% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend0.R
\name{extend0}
\alias{extend0}
\title{Turn spiketimes into a signal.}
\usage{
extend0(v, by = 1, bounds = NA, center = TRUE, vectoronly = FALSE, neg = FALSE)
}
\arguments{
\item{v}{A vector of spiketimes.}

\item{by}{Spacing of observations in the series. Defaults to 1.}

\item{bounds}{Optional. If no value is provided a signal between the first and last value of v is generated. One value = lower bound, two values = lower and upper bound.}

\item{center}{Should the time series be centered around 0? Defaults to \code{TRUE}}

\item{vectoronly}{If \code{TRUE}, returns only the time signal vector. (defaults to \code{FALSE})}
}
\value{
A dt with 2 columns. V1 = time, V2 = signal. If \code{vectoronly = TRUE} returns only the signal (defaults to \code{FALSE}).
}
\description{
The signal goes high on times with a spike. 0 = no spike, 1 = spike if not centered. For internal use.
}
\details{
Spiketimes should either be in the original 30000Hz or be downsampled to even chunks such as .5ms for example. This should be specified in the \code{by} parameter. Ie if the spikes are sampled at a rate of .5ms (0, 0.5, 1, 1.5...) \code{by} should be .5.
}
