% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fftspikes.R
\name{fftspikes}
\alias{fftspikes}
\title{Apply an FFT to phyoutput}
\usage{
fftspikes(x, clstr, len = 30000, sample.frq = 30000, tpr = NA, pwronly = FALSE)
}
\arguments{
\item{x}{A phyouput object}

\item{clstr}{What cluster to apply fft to}

\item{len}{Length of each window in the same unit as the timeseries in \code{x}. (Ie if x is in the original 30 000Hz setting len to 30000 will make windows 1 second long)}

\item{sample.frq}{The frequency at which the time series was sampled}

\item{tpr}{Taper to apply has to be the same length as windw}

\item{pwronly}{If \code{FALSE} (default) returns the DFT, PSD and frequencies.}
}
\value{
A dt with raw DFT results from welch's method, PSD and frequency bins.
}
\description{
Apply an FFT to phyoutput
}
