% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISIdistr.R
\name{ISIdistr}
\alias{ISIdistr}
\title{Compute distribution of ISI}
\usage{
ISIdistr(x, binsize, extend = TRUE, relative = TRUE, minisi = 0, maxisi = 500)
}
\arguments{
\item{x}{A phyoutput with original spiketimes}

\item{binsize}{Binsize in ms}

\item{extend}{Should "missing" bins (ie those representing 0 ISI) be added?}

\item{relative}{Should the relative ISI of each cluster be included?}

\item{minisi}{The minimum ISI to include when adding "missing" bins.}

\item{maxisi}{The maximum ISI to include when adding "missing" bins.}
}
\value{
A dt with the distribution of ISI by cluster
}
\description{
Compute distribution of ISI
}
