% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadSpikes.R
\name{loadSpikes}
\alias{loadSpikes}
\title{Title}
\usage{
loadSpikes(path, triggerfile)
}
\arguments{
\item{path}{Path to ohy output}

\item{triggerfile}{Name of triggerchannel, has to be in the 'path' dir and in .csv form.}
}
\value{
A list
}
\description{
Title
}
\examples{
loadSpikes(path = "C:/Users/Elias/Desktop/phyish/Phyr/Våra", triggerfile = "asd.csv")
}
