% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plothz.R
\name{plothz}
\alias{plothz}
\title{Plot the binned frequency of a cluster}
\usage{
plothz(x, clstrs, t = "full", plotline = TRUE)
}
\arguments{
\item{x}{A phyoutput from spikesper}

\item{clstrs}{Clusters to include in plot}

\item{t}{A numeric vector of size 2 containing desired timespan.}

\item{plotline}{Should the raw frequency be plotted below the smoothed?}
}
\value{
A plot of smoothed frequency.
}
\description{
The phyoutput used should be the result from \link{spikesper}
}
