% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotpwr.R
\name{plotpwr}
\alias{plotpwr}
\title{PSD plot for phyoutput}
\usage{
plotpwr(x, hzlim = 500)
}
\arguments{
\item{x}{An object created by \code{\link{fftspikes}}.}

\item{hzlim}{The max frequency to include in the plot.}
}
\value{
A ggplot object showing the power spectral density of a cluster.
}
\description{
PSD plot for phyoutput
}
