% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotraster.R
\name{plotraster}
\alias{plotraster}
\title{Plot rasterplot}
\usage{
plotraster(x, clstr, timeconv = 30, CS = 0, US = 300)
}
\arguments{
\item{x}{An object with the attribute \code{triggered} (see \link{triggers()})}

\item{clstr}{Cluster id to plot}
}
\value{
A rasterplot of spikes in relation to triggers
}
\description{
Plot rasterplot
}
