% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotrelhz.R
\name{plotrelhz}
\alias{plotrelhz}
\title{Title}
\usage{
plotrelhz(
  x,
  clstrs = NA,
  by = c("cluster", "trigger", "depthraster", "depthline"),
  CS = 0,
  US = 300,
  abs = FALSE
)
}
\arguments{
\item{abs}{}
}
\description{
Title
}
