% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rel_response.R
\name{rel_response}
\alias{rel_response}
\title{Calculate relative responses to triggers}
\usage{
rel_response(x, period = 30, depthdiv = NA)
}
\arguments{
\item{x}{An object produced by \code{triggers()}}

\item{period}{The period in ms by which to summarize}

\item{depthdiv}{Should the depths be summarized into equally sized ranges? depthdiv = NA -> don't summarize(default). depthdiv = 4 -> divide into 4 depth ranges and calculate mean at each bin.}
}
\value{
A phyoutput object containing:
\itemize{
\item \code{spiketimes} - Binned spike frequencies, baseline frequencies by cluster, trigger and depth as well as relative frequency of each bin compared to baseline.
\item \code{info} - General info about each cluster from phy
\item \code{clustermeans} - Relative and absolute Hz across trigger sessions by \strong{cluster} and time.
\item \code{depthmeans} - Relative and absolute Hz across trigger sessions by \strong{depth} and time
}
}
\description{
Responses are summarized by cluster, trigger session and depth. A baseline is calculated as the number of spikes before CS / time included before CS. Each trigger session is given a separate baseline.
When a short \code{period} is used a cluster with a low firerate may have bins with higher frequency than true. I.e when a period of 10ms is used, 1 spike in 10 ms translates to 100 spikes/s for that bin. This is usually "fixed" whem averaging across trigger sessions as many bins will be 0.
}
\seealso{
\code{\link[=triggers]{triggers()}} \code{\link[=plotrelhz]{plotrelhz()}}
}
