% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spikeISI.R
\name{spikeISI}
\alias{spikeISI}
\title{Calculate ISI}
\usage{
spikeISI(x, ISIonly = FALSE)
}
\arguments{
\item{x}{A list-object created by \code{\link{loadSpikes}}}

\item{ISIonly}{Should a dt containing only cluster and ISI be returned? Defaults to FALSE.}
}
\value{
ISIonly = FALSE -> Modifies the list-object adding a column to spiketimes with ISI.
ISIonly = TRUE -> A dt with cluster and ISI for each spike.
}
\description{
Calculate ISI
}
\examples{
dt <- loadSpikes(path, triggerfile)
spikeISI(dt)
}
