% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spikesper.R
\name{spikesper}
\alias{spikesper}
\title{Create dt with spikes/period}
\usage{
spikesper(x, period, full = TRUE)
}
\arguments{
\item{x}{A list object created by \code{\link{loadSpikes}}.}

\item{period}{The period in milliseconds by which to summarize.}

\item{full}{Should a "full" time series including periods without spikes be returned? Defaults to \code{TRUE}.}
}
\value{
A list object with spikes per period as specified by \code{per}. Includes periods with no spikes by default.
}
\description{
Summarizes the spiketimes to spikes/period.
Removes the attribute "\code{ogspiketimes}" and adds the specified \code{period} as an attribute.
}
\examples{
# Get spikes per second
dt <- loadSpikes(path, triggerfile)
dt <- spikesper(dt, 1)
# Get spikes per 10 seconds
dt <- spikesper(dt, 10)
}
