% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triggers.R
\name{triggers}
\alias{triggers}
\title{Mark spikes by their nearest trigger}
\usage{
triggers(x, back = 500, forward = 600)
}
\arguments{
\item{x}{A list object as created by \code{\link{loadSpikes}}}

\item{back}{The time in ms to look backwards from trigger. Defaults to 500.}

\item{forward}{The time in ms to look forwards from trigger. Defaults to 600.}
}
\value{
The same list object but the "spiketimes" dt has an additional column "ntrig"
}
\description{
A function which adds a column to \code{x} marking its nearest trigger as specified with the \code{forward} and \code{back} params.
}
\examples{
dt <- loadSpikes(path, triggerfile)
triggers(dt)
}
