% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtextend.R
\name{dtextend}
\alias{dtextend}
\title{Title}
\usage{
dtextend(dt, col, group, step, fill = 0, min = 0, max = "auto", ask = TRUE)
}
\arguments{
\item{dt}{A dt}

\item{col}{Quoted name of column to expand by}

\item{group}{Quoted name of column to group by}

\item{step}{Steps by which to pad the column specified in \code{col}}

\item{fill}{Value to fill other columns with}

\item{min}{The minimum value \code{col} should take}

\item{max}{The maximum value \code{col} should take. If none is specified the max value of \code{col} will be used. If the number of rows in the final dt will surpass 1 000 000 the user will be given the option to provide a max in terminal if \code{ask = TRUE}.}

\item{ask}{Should the user be asked to provide a max if final dt will be > 1 000 000 long?}
}
\value{
A dt with all combinations if \code{col} & \code{group} with restrictions as described above. All columns in new rows other than \code{col} & \code{group} will be filled with 0 unless specified otherwise.
}
\description{
Title
}
