% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadSpikes.R
\name{loadSpikes}
\alias{loadSpikes}
\title{Load spikes from Phy output}
\usage{
loadSpikes(path, triggerfile = NA, minfr = 0)
}
\arguments{
\item{path}{Path to phy output}

\item{triggerfile}{Name of triggerchannel, has to be in the 'path' dir and in .csv form.}

\item{minfr}{The minimum mean firerate to include}
}
\value{
An list-object with classes \code{phyoutput} and \code{ogspiketimes} containing:
\itemize{
\item "spiketimes", a dt with spiketimes, their cluster, channel and depth.
\item "triggers", a vector of triggertimes
\item "info", a dt with info from phy on each cluster
}

All spike-/trigger times are represented in their original 30 000Hz form.
Associated attributes of importance:
\itemize{
\item \code{class} -> ogspiketimes - denotes original spiketimes without summarizing into bins or similar
\item \code{clusters}: A chr vector of all clusters in contained
\item \code{tfactorms}: The factor by which to divide the time column to convert it to ms.
}
}
\description{
Returns a phyoutput containing every spiketime and associated info marked as "good" in Phy.\cr
Associated info includes depth, channel, triggertimes (if specified in \code{triggerfile}), amplitude, and mean firerate
}
\examples{
loadSpikes(path = "C:/Users/Elias/Desktop/phyish/Phyr/Våra", triggerfile = "asd.csv")
}
