% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadSpikes.R
\name{loadSpikes}
\alias{loadSpikes}
\title{Load spikes from Phy output}
\usage{
loadSpikes(path, triggerfile = NA)
}
\arguments{
\item{path}{Path to phy output}

\item{triggerfile}{Name of triggerchannel, has to be in the 'path' dir and in .csv form.}
}
\value{
A list containing:
\itemize{
\item "spiketimes", a dt with spiketimes, their cluster, channel and depth.
\item "triggers", a vector of triggertimes
\item "info", a dt with info from phy on each cluster
}

All spike-/trigger times are represented in their original 30 000Hz form.
}
\description{
loadSpikes returns a list object containing every spiketime and associated info marked as "good" in Phy.\cr
Associated info includes depth, channel, triggertimes, amplitude, and mean firerate
}
\examples{
loadSpikes(path = "C:/Users/Elias/Desktop/phyish/Phyr/Våra", triggerfile = "asd.csv")
}
