% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadSpikes.R
\name{loadSpikes}
\alias{loadSpikes}
\title{Title}
\usage{
loadSpikes(path, triggerfile)
}
\arguments{
\item{path}{Path to phy output}

\item{triggerfile}{Name of triggerchannel, has to be in the 'path' dir and in .csv form.}
}
\value{
A list containing:
\itemize{
\item "spiketimes", a dt with spiketimes, their cluster, channel and depth.
\item "triggers", a vector of triggertimes
\item "info", a dt with info from phy on each cluster
}

All spike-/trigger times are represented in their original 30 000Hz form.
}
\description{
Title
}
\examples{
loadSpikes(path = "C:/Users/Elias/Desktop/phyish/Phyr/Våra", triggerfile = "asd.csv")
}
