% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotrelhz.R
\name{plotrelhz}
\alias{plotrelhz}
\title{Title}
\usage{
plotrelhz(
  x,
  by = c("cluster", "trigger", "depthraster", "depthline"),
  clstrs = NA,
  CS = 0,
  US = 300,
  abs = FALSE
)
}
\arguments{
\item{x}{An object returned by \link{\code{rel_response}}}

\item{abs}{}
}
\description{
Title
}
