% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spikesper.R
\name{spikesper}
\alias{spikesper}
\title{Create dt with spikes/period}
\usage{
spikesper(x, period)
}
\arguments{
\item{x}{A list object created by \code{\link{loadSpikes}}.}

\item{period}{The period in seconds by which to summarize.}
}
\value{
A list object with spikes per period as specified by \code{per}.
}
\description{
Create dt with spikes/period
}
\examples{
# Get spikes per second
dt <- loadSpikes(path, triggerfile)
dt <- spikesper(dt, 1)
# Get spikes per 10 seconds
dt <- spikesper(dt, 10)
}
