% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triggers.R
\name{triggers}
\alias{triggers}
\title{Mark spikes by their nearest trigger}
\usage{
triggers(x, back = 500, forward = 600)
}
\arguments{
\item{x}{A list object as created by \code{\link{loadSpikes}}}

\item{back}{The time in ms to look backwards from trigger. Defaults to 500.}

\item{forward}{The time in ms to look forwards from trigger. Defaults to 600.}
}
\value{
A new phyouput object containing info and spikes near triggers.
}
\description{
A function which creates a new \code{phyoutput} object only containing spikes near a trigger (as specified by \code{back} & \code{forward}).
}
\examples{
dt <- loadSpikes(path, triggerfile)
newdt <- triggers(dt)
}
